+++ webview-ui/src/components/chat/ChatRow.tsx
@@ -1024,7 +1024,7 @@ export const ChatRowContent = ({ message, isExpanded, onToggleExpand, lastModifi
 								</div>
 								<div>
 									Cline tried to access <code>{message.text}</code> which is blocked by the{" "}
-									<code>.clineignore</code>
+									<code>.ogignore</code>
 									file.
 								</div>
 							</div>